DASH-354/refactor: repository 구조 변경
