+++ src/main/java/io/saim/dash/coupon/common/repository/jpa/IssueRequestJpaRepository.java
@@ -1,4 +1,4 @@
-package io.saim.dash.coupon.common.repository.IssueRequest;
+package io.saim.dash.coupon.common.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
